requirements.txt
Pillow==12.0.0
PyQt5==5.15.11